move simplecsv to libraries
